+++ universe/index_test.go
@@ -230,7 +230,7 @@ var _ = Describe("Index Universe", func() {
 	})
 
 	Describe("SP500 and Nasdaq100 convenience constructors", func() {
-		It("SP500 returns a universe that queries 'sp500'", func() {
+		It("SP500 returns a universe that queries 'SPX'", func() {
 			provider := &mockIndexProvider{
 				assetResults: map[int64][]asset.Asset{
 					now.Unix(): {aapl},
@@ -244,7 +244,7 @@ var _ = Describe("Index Universe", func() {
 			Expect(assets).To(ConsistOf(aapl))
 		})
 
-		It("Nasdaq100 returns a universe that queries 'ndx100'", func() {
+		It("Nasdaq100 returns a universe that queries 'NDX'", func() {
 			provider := &mockIndexProvider{
 				assetResults: map[int64][]asset.Asset{
 					now.Unix(): {goog},